# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Figma Challenge",
        "description": "Documentation des schémas des requêtes",
        "contact": {
            "name": "Louqmane Inoussa"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/notes": {
            "get": {
                "tags": [
                    "Notes Jury"
                ],
                "summary": "Lister toutes les notes de jury",
                "operationId": "625a58996d697c0651d8aa409724c1aa",
                "responses": {
                    "200": {
                        "description": "Liste des notes"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notes Jury"
                ],
                "summary": "Attribuer une note",
                "operationId": "cc6aca6ed6e3079e4618e5d513d9d1a5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "user_id",
                                    "soumission_id",
                                    "graphisme",
                                    "animation",
                                    "navigation"
                                ],
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "soumission_id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "graphisme": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "animation": {
                                        "type": "integer",
                                        "example": 8
                                    },
                                    "navigation": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "commentaire": {
                                        "type": "string",
                                        "example": "Très bon travail globalement"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note enregistrée"
                    }
                }
            }
        },
        "/api/notes/{id}": {
            "get": {
                "tags": [
                    "Notes Jury"
                ],
                "summary": "Afficher une note",
                "operationId": "a543dbddf0b155f408e1f047ff5b8592",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la note"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notes Jury"
                ],
                "summary": "Supprimer une note",
                "operationId": "4e837ec1c5e2ed6f00134ea84f83d89c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note supprimée"
                    }
                }
            }
        },
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Lister toutes les notifications",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "responses": {
                    "200": {
                        "description": "Liste des notifications"
                    }
                }
            },
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Créer une notification",
                "operationId": "91f8e779a9e46cb5cdda969e5a928a94",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "content",
                                    "audience"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Deadline demain !"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "N'oubliez pas de soumettre votre projet avant minuit."
                                    },
                                    "audience": {
                                        "type": "string",
                                        "example": "challengers"
                                    },
                                    "scheduled_at": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-06-24T12:00:00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Notification créée"
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Afficher une notification",
                "operationId": "f2edf725f41dd6c9a9338a6a7b39d779",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail de la notification"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Supprimer une notification",
                "operationId": "54c4eb5ec5dc0d1e2916cec2fced1056",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notification supprimée"
                    }
                }
            }
        },
        "/api/projects": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Lister tous les projets",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "responses": {
                    "200": {
                        "description": "Liste des projets"
                    }
                }
            },
            "post": {
                "tags": [
                    "Projects"
                ],
                "summary": "Créer un nouveau projet",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "category",
                                    "challenge_id",
                                    "deadline"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Repenser une plateforme de transport"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Créer une expérience fluide pour les utilisateurs"
                                    },
                                    "cover": {
                                        "type": "string",
                                        "example": "project.png"
                                    },
                                    "challenge_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "category": {
                                        "type": "string",
                                        "example": "transport"
                                    },
                                    "start_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-10"
                                    },
                                    "deadline": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2025-06-20"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Projet créé"
                    }
                }
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Projects"
                ],
                "summary": "Afficher un projet",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet détaillé"
                    }
                }
            },
            "put": {
                "tags": [
                    "Projects"
                ],
                "summary": "Mettre à jour un projet",
                "operationId": "491d11cc6efd20335421a859159e2487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Projet mis à jour"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Projects"
                ],
                "summary": "Supprimer un projet",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Projet supprimé"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "StoreProjectRequest": {
                "required": [
                    "title",
                    "category",
                    "challenge_id",
                    "deadline"
                ],
                "properties": {
                    "title": {
                        "type": "string",
                        "example": "Projet Mobilité"
                    },
                    "description": {
                        "type": "string",
                        "example": "Une plateforme pour organiser le transport urbain"
                    },
                    "category": {
                        "type": "string",
                        "example": "transport"
                    },
                    "cover": {
                        "type": "string",
                        "example": "project-cover.png"
                    },
                    "challenge_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-01"
                    },
                    "deadline": {
                        "type": "string",
                        "format": "date",
                        "example": "2025-07-31"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}